Redshift sampling from Schechter luminosity function (#50)
* Functions to sample redshifts from the Schechter luminosity function according to the model of Herbel et al. (2017)

* Functions to sample redshifts from the Schechter luminosity function according to the model of Herbel et al. (2017)

* Created astronomy.py and included the conversion from absolute magnitudes to luminosities and its inverse.

* Included comoving element. No analytical CDF anymore so I use the un-normalised pdf.

* The 'models.py' function was deleted as we don't wanna have hardcoded parameters.

The '_rescale_luminosity_function' was not longer required because the function changed with the last commits.

The test were adapted to the new syntax.

* Some bug fixes because the some tests failed.

* Some cosmetic changes plus adjustments for the cosmology instance

Cosmology as well as resolution are arguments now.
Cosmology has to be an instance of an astropy.cosmology class.
Resolution gives the length of the redshift parameter which is later be used in the interpolation. Some tests showed that 100 is a reasonable default.